Remove: ModifyProfile를 profile에 병합
